api filters for renamed packages
