+++ ac.soton.coda.simulator.ui/src/ac/soton/coda/internal/simulator/handlers/StartSimAction.java
