+++ ac.soton.coda.simulator.ui/src/ac/soton/coda/simulator/Activator.java
@@ -9,7 +9,7 @@
  * Contributors:
  *      University of Southampton - Initial API and implementation
  *******************************************************************************/
-package org.coda.simulator.ui;
+package ac.soton.coda.simulator;
 
 import org.eclipse.jface.resource.ImageDescriptor;
 import org.eclipse.ui.plugin.AbstractUIPlugin;
@@ -21,7 +21,7 @@ import org.osgi.framework.BundleContext;
 public class Activator extends AbstractUIPlugin {
 
 	// The plug-in ID
-	public static final String PLUGIN_ID = "org.coda.simulator.ui"; //$NON-NLS-1$
+	public static final String PLUGIN_ID = "ac.soton.coda.simulator"; //$NON-NLS-1$
 
 	// The shared instance
 	private static Activator plugin;
